fix ci build

+++ src/nvim/api/vim.c
@@ -789,7 +789,7 @@ Integer nvim_echo(ArrayOf(Tuple(String, *HLGroupID)) chunks, Boolean history, Di
     return 0;
   }
 
-  MsgID id = msg_multihl(opts->id, hl_msg, kind, history, opts->err, status.data, opts->percentage);
+  MsgID id = msg_multihl(opts->id, hl_msg, kind, history, opts->err, status.data, (int)opts->percentage);
 
   if (opts->verbose) {
     verbose_leave();
